refactor: extract discovery output helpers to cli/commands/discovery.py
Move 11 discovery-internal symbols (4 exception classes, 1 compiled regex,
6 helper functions) from generator.py into the new canonical module
src/cja_auto_sdr/cli/commands/discovery.py. generator.py re-imports all
moved symbols for backwards compatibility. Adds test_discovery_extraction.py
to verify the new module is importable with the expected public surface.
